+++ eslint.config.ts
@@ -1,0 +1,9 @@
+import neolutionEslintConfig from "@neolution-ch/eslint-config-neolution";
+
+export default [
+  ...neolutionEslintConfig.configs.flat.getConfig({
+    ...neolutionEslintConfig.configs.flat.defaults.typescript,
+    jest: true,
+    jsdocRequireJsdoc: true,
+  }),
+];