+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Rollup and Jest configs in TypeScript
+- Ignore code coverage output folder
+- Migrate to eslint 9 flat config
 
 ### Removed
 
@@ -24,7 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Split up `CI` workflow into two separate workflows for the comment posting to work for PRs from forks.
-- Ignore code coverage output folder
 
 ### Fixed